backend/requirements.txt
# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# OCR
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# Deep Learning Inference
onnxruntime>=1.15.0

# Image Inpainting
simple-lama-inpainting>=0.1.2

# Utilities
requests>=2.31.0
